setup the Bot for Work
