+++ .gitIgnore
@@ -1,1 +1,3 @@
 config.py
+/.idea/
+birthdays.json
